# [Serverless-Init] Improve Tags Performance