--- a/pkg/config/env/environment.go
+++ b/pkg/config/env/environment.go
@@ -81,7 +81,7 @@ func IsHostSysAvailable() bool {
 	return true
 }
 
-// IsServerless returns whether the Agent is running in a Lambda function
-func IsServerless() bool {
+// IsLambda returns whether the Agent is running in a Lambda function
+func IsLambda() bool {
 	return os.Getenv("AWS_LAMBDA_FUNCTION_NAME") != ""
 }