--- a/releasenotes/notes/serverless-init-cache-tags-to-string-1f3c0f0f5090c242.yaml
+++ b/releasenotes/notes/serverless-init-cache-tags-to-string-1f3c0f0f5090c242.yaml
@@ -1,0 +1,4 @@
+---
+enhancements:
+  - |
+    Cache result of `TagsToString()` in serverless-init to improve CPU performance.