### What does this PR do?

[Serverless-init](https://hub.docker.com/r/datadog/serverless-init) is a special build of the agent for Serverless containers like Google Cloud Run or Azure Container Apps.
- Create a custom JSON encoder for serverless-init
- Cache stringified tags since tags are constant in serverless-init. There is only one log source in serverless-init
- This significantly improves CPU performance. In my tests, we eliminated ~7% CPU usage that was being wasted:
```
% go tool pprof ./saved-profiles/cpu-tagstostring-before.pb.gz 
Duration: 60.14s, Total samples = 8.25s (13.72%)
(pprof) top -cum TagsToString
Showing top 10 nodes out of 52
      flat  flat%   sum%        cum   cum%
         0     0%     0%      0.60s  7.27%  github.com/DataDog/datadog-agent/pkg/logs/message.(*MessageMetadata).TagsToString (inline)
         0     0%     0%      0.60s  7.27%  github.com/DataDog/datadog-agent/pkg/logs/message.(*Origin).TagsToString
         0     0%     0%      0.60s  7.27%  github.com/DataDog/datadog-agent/pkg/logs/processor.(*Processor).processMessage
         0     0%     0%      0.60s  7.27%  github.com/DataDog/datadog-agent/pkg/logs/processor.(*Processor).run
         0     0%     0%      0.60s  7.27%  github.com/DataDog/datadog-agent/pkg/logs/processor.(*jsonServerlessEncoder).Encode

% go tool pprof ./saved-profiles/cpu-tagstostring-after.pb.gz
Duration: 60.08s, Total samples = 7.93s (13.20%)
(pprof) top -cum TagsToString
Focus expression matched no samples
```

Flame graph before:
<img width="1000" height="496" alt="Screenshot 2025-10-03 at 11 33 18 AM" src="https://github.com/user-attachments/assets/2c961269-a5b9-455e-a29a-3bff338b58f7" />


You can view the profiles yourself: [before](https://github.com/DataDog/serverless-init-performance-testing/blob/main/saved-profiles/cpu-tagstostring-before.pb.gz) and [after](https://github.com/DataDog/serverless-init-performance-testing/blob/main/saved-profiles/cpu-tagstostring-after.pb.gz)

### Motivation

Improve serverless-init performance

https://datadoghq.atlassian.net/browse/SVLS-7657

### Describe how you validated your changes

Via profiles and manually. I verified that tags still get added as expected on Google Cloud Run.

We also have [e2e tests](https://github.com/DataDog/serverless-e2e-serverless-init-gcp-tests) and [long-running tests](https://github.com/DataDog/serverless-init-self-monitoring) to detect regressions. We always run them before doing a release.

### Additional Notes